+++ changelog.md
@@ -1,6 +1,18 @@
 # Better Deep Dark
 ## Development Changelog
 
+## 2024.07.06:
+- Echo Stick, Soul Piece, Echo Sword, Echo Pickaxe, Echo Axe, Echo Shovel, Echo Hoe, Echo Helmet, Echo Chestplate, Echo Leggings, and Echo Boots are now fire-resistant.
+- Removed Knowledge Soup.
+- The Echo Upgrade Smithing Template now works exactly as a Netherite Upgrade Smithing Template.
+- Renamed Infected Warden's Horn to Slimed Warden's Horn.
+- Slimed Warden's Horn and Warden's Horn now appear first in the game's creative tab.
+- The Warden's Horn and Slimed Warden's Horn now give two effects, one is a positive effect, and the second is negative.
+- Slimed Warden's Horn can now be made using a Warden's Horn and a Slime Ball.
+- The Warden now has a chance to drop Warden's Horn of 5%.
+- Ancient city chests can now contain Warden's Horn and Slimed Warden's Horn.
+- Slimed Warden's Horn, Warden's Horn, and Soul Piece are now stackable to 16.
+
 ## 2024.07.04 (3):
 - Echo armor is now obtainable by using the smithing table.
 - Ancient city ice box chest will now contain Soul Piece, Echo Stone, and Echo Stick.